+++ public/pages/CreateDetector/components/DefineDetector/components/DetectorDataSource/DetectorDataSource.tsx
@@ -5,20 +5,31 @@
 
 import React, { Component } from 'react';
 import { ContentPanel } from '../../../../../../components/ContentPanel';
-import { EuiComboBox, EuiComboBoxOptionOption, EuiFormRow, EuiSpacer } from '@elastic/eui';
+import {
+  EuiComboBox,
+  EuiComboBoxOptionOption,
+  EuiFormRow,
+  EuiSpacer,
+  EuiCallOut,
+  EuiTextColor,
+} from '@elastic/eui';
 import { FormFieldHeader } from '../../../../../../components/FormFieldHeader/FormFieldHeader';
 import { IndexOption } from '../../../../../Detectors/models/interfaces';
 import { MIN_NUM_DATA_SOURCES } from '../../../../../Detectors/utils/constants';
 import IndexService from '../../../../../../services/IndexService';
 import { NotificationsStart } from 'opensearch-dashboards/public';
 import { errorNotificationToast } from '../../../../../../utils/helpers';
+import _ from 'lodash';
+import { FieldMappingService } from '../../../../../../services';
 
 interface DetectorDataSourceProps {
   detectorIndices: string[];
   indexService: IndexService;
+  filedMappingService: FieldMappingService;
   isEdit: boolean;
   onDetectorInputIndicesChange: (selectedOptions: EuiComboBoxOptionOption<string>[]) => void;
   notifications: NotificationsStart;
+  detector_type: string;
 }
 
 interface DetectorDataSourceState {
@@ -26,18 +37,22 @@ interface DetectorDataSourceState {
   fieldTouched: boolean;
   indexOptions: IndexOption[];
   errorMessage?: string;
+  message: string[];
 }
 
 export default class DetectorDataSource extends Component<
   DetectorDataSourceProps,
   DetectorDataSourceState
 > {
+  private indicesMappings: any = {};
+
   constructor(props: DetectorDataSourceProps) {
     super(props);
     this.state = {
       loading: true,
       fieldTouched: props.isEdit,
       indexOptions: [],
+      message: [],
     };
   }
 
@@ -82,17 +97,74 @@ export default class DetectorDataSource extends Component<
     this.onSelectionChange(parsedOptions);
   };
 
-  onSelectionChange = (options: EuiComboBoxOptionOption<string>[]) => {
+  onSelectionChange = async (options: EuiComboBoxOptionOption<string>[]) => {
+    const allIndices = _.map(options, 'label');
+    for (let indexName in this.indicesMappings) {
+      if (allIndices.indexOf(indexName) === -1) {
+        // cleanup removed indexes
+        delete this.indicesMappings[indexName];
+      }
+    }
+
+    for (const indexName of allIndices) {
+      if (!this.indicesMappings[indexName]) {
+        const detectorType = this.props.detector_type.toLowerCase();
+        const result = await this.props.filedMappingService.getMappingsView(
+          indexName,
+          detectorType
+        );
+        result.ok && (this.indicesMappings[indexName] = result.response.unmapped_field_aliases);
+      }
+    }
+
+    if (!_.isEmpty(this.indicesMappings)) {
+      let firstMapping: string[] = [];
+      let firstMatchMappingIndex: string = '';
+      let message: string[] = [];
+      for (let indexName in this.indicesMappings) {
+        if (this.indicesMappings.hasOwnProperty(indexName)) {
+          if (!firstMapping.length) firstMapping = this.indicesMappings[indexName];
+          !firstMatchMappingIndex.length && (firstMatchMappingIndex = indexName);
+          if (!_.isEqual(firstMapping, this.indicesMappings[indexName])) {
+            message = [
+              `The below log sources don't have the same fields, please consider creating separate detectors for them.`,
+              firstMatchMappingIndex,
+              indexName,
+            ];
+            break;
+          }
+        }
+      }
+
+      this.setState({ message });
+    }
+
     this.props.onDetectorInputIndicesChange(options);
   };
 
   render() {
     const { detectorIndices } = this.props;
-    const { loading, fieldTouched, indexOptions, errorMessage } = this.state;
+    const { loading, fieldTouched, indexOptions, errorMessage, message } = this.state;
     const isInvalid = fieldTouched && detectorIndices.length < MIN_NUM_DATA_SOURCES;
     return (
       <ContentPanel title={'Data source'} titleSize={'m'}>
         <EuiSpacer size={'m'} />
+        {message.length ? (
+          <>
+            <EuiCallOut title="Detector configuration warning" color="warning" iconType="alert">
+              {message.map((messageItem: string, index: number) => (
+                <EuiTextColor
+                  color={index === 0 ? 'default' : 'warning'}
+                  key={`callout-message-part-${index}`}
+                >
+                  {messageItem}
+                  <br />
+                </EuiTextColor>
+              ))}
+            </EuiCallOut>
+            <EuiSpacer size={'m'} />
+          </>
+        ) : null}
         <EuiFormRow
           label={
             <FormFieldHeader headerTitle={'Select or input source indexes or index patterns'} />
